% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_vmeta_num_class.R
\name{filter_vmeta_num}
\alias{filter_vmeta_num}
\title{Filter by numeric variable metadata}
\usage{
filter_vmeta_num(mode = "exact", level, factor_name, ...)
}
\arguments{
\item{mode}{(character) Mode of action. Allowed values are limited to the following: \itemize{\item{\code{"exact"}: Samples with exact numeric value are retained.}\item{\code{"above"}: Samples with numeric values above the threshold are retained.}\item{\code{"below"}: Samples with numeric values below the threshold are retained.}} The default is \code{"exact"}.}

\item{level}{(name)}

\item{factor_name}{(character) The factor name for filtering.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{filter_vmeta_num} object with the following \code{output} slots:
\tabular{ll}{
\code{filtered} \tab          (DatasetExperiment)  \cr
}
}
\description{
The data is filtered based on numeric variable metadata using the specified mode and level.
}
\examples{
D = MTBLS79_DatasetExperiment()
M = filter_vmeta_num(mode='above', level=10, factor_name='QC')
M = model_apply(M, D)
}
