% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_charts.R
\name{pls_regcoeff_plot}
\alias{pls_regcoeff_plot}
\title{pls_regcoeff_plot class}
\usage{
pls_regcoeff_plot(ycol = 1, ...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Plots the regression coefficients of a PLSDA model.
}
\examples{
D = iris_DatasetExperiment()
M = mean_centre()+PLSDA(factor_name='Species')
M = model_apply(M,D)

C = pls_regcoeff_plot(ycol='setosa')
chart_plot(C,M[2])
}
\references{
Liland K, Mevik B, Wehrens R (2022).
\emph{pls: Partial Least Squares and Principal Component Regression}.
R package version 2.8-1, \url{https://CRAN.R-project.org/package=pls}.

Wickham H (2016).
\emph{ggplot2: Elegant Graphics for Data Analysis}.
Springer-Verlag New York.
ISBN 978-3-319-24277-4, \url{https://ggplot2.tidyverse.org}.
}
