% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPPstructToolbox.R
\docType{package}
\name{MAPPstructToolbox}
\alias{MAPPstructToolbox}
\title{MAPPstructToolbox: Examples of tools built using the Statistics in R Using Class Templates (struct) package}
\description{
This package extends the classes defined in the struct package
}
